LayoutTests/webrtc/datachannel/bufferedAmountLowThreshold-expected.txt

CONSOLE MESSAGE: line 76: channel send is throwing
CONSOLE MESSAGE: line 76: channel send is throwing

PASS Default buffer threshold 
PASS Large buffer threshold reached 
PASS Medium buffer threshold not reached 
